refactor: move document utils
